Revert "Revert "Ready for submission""
This reverts commit 683da6d0cfeea250e83b46029f98bf4ff6dfca23.

+++ Procedures/DeleteStudent.php
@@ -1,4 +1,5 @@
 <?php
+$num_rows = 0;
 
 if($_POST){
 	$con = mysqli_connect("localhost", "root", "", "StudentSystem"); //Username and Password to connect to phpMyAdmin
@@ -6,37 +7,40 @@ if($_POST){
 	    die('Could not connect: ' . mysql_error());
 	}
 
-	//mysqli_select_db("StudentSystem", $con);
-
 	$ID = $_POST['ID'];
 	$FirstName = $_POST['FirstName'];
 	$LastName = $_POST['LastName'];
 
 	$ID = htmlspecialchars($ID);
 	$FirstName = htmlspecialchars($FirstName);
-	$LastName = htmlspecialchars($LastName);	
+	$LastName = htmlspecialchars($LastName);
 
+	if($ID == "" || $FirstName == "" || $LastName == ""){
+		die('One of the fields are missing');
+	}
 
+	$SQL = "SELECT * FROM STUDENT WHERE ID = '$ID' and FirstName = '$FirstName' and LastName = '$LastName';";
+	$result = mysqli_query($con,$SQL);
+	$num_rows = mysqli_num_rows($result);
 
-	$SQL = "SELECT * FROM STUDENT WHERE ID ='$ID' and FirstName ='$FirstName' and LastName = '$LastName';";
-	$result = mysqli_query($con, $SQL);
-	$num_row = 0;
+	if($num_rows >= 1){
 
-	if($num_row > 0){
 	$deleteStudent = "
 	DELETE From Student Where ID='$ID' and FirstName='$FirstName' and LastName='$LastName';";
-
-	if (!mysqli_query($con,$deleteStudent)) {
+	
+	if (!mysqli_query($con,$deleteStudent)) 
+	{
             die('Error: ' . mysqli_error($con));
-        }
+    }
 	echo "<h1>Deleted Student Record!</h2>";
-
 	mysqli_close($con);
 	}
-	else{
-		die('Error: Student Not Found');
+	else
+	{
+	die('Error: No such student');
 	}
-
+	
 }
+
 
 ?>
